pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "medvision-ai"
version = "1.0.0"
description = "AI-assisted medical image analysis — X-ray, MRI, dermoscopy"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "AXONN AI STUDIO" }]
requires-python = ">=3.10"
keywords = ["medical imaging", "deep learning", "radiology", "dermatology", "explainable AI"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "gradio>=4.0.0",
    "shap>=0.44.0",
    "pyyaml>=6.0",
    "huggingface-hub>=0.22.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/your-username/medvision-ai"
Documentation = "https://github.com/your-username/medvision-ai/tree/main/docs"
"Bug Tracker" = "https://github.com/your-username/medvision-ai/issues"
"HuggingFace" = "https://huggingface.co/your-username/medvision-ai"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
